Initialize agents on test setup to ensure schema created
message.test.js was failing because createMessage was being called before the agent schema was initialized (specifically the fetchThread method of messageService throws an error when it tries to populate the 'agents' field of a Thread). This doesn't happen in a server env because initializeAgents is called by setup when server starts.